requirements.txt
pygithub~=1.53
pathlib~=1.0.1
progressbar2
pygit2
numpy~=1.19.1
pyvis